refactor: move Circuit structs to primitives & implement ReadSFX
