+++ primitives/src/circuit/mod.rs
@@ -1,0 +1,5 @@
+pub mod traits;
+pub mod types;
+
+pub use traits::*;
+pub use types::*;
